Move DevinLogger.kt to devin-api module
