--- devin-no-op/src/main/java/com/khosravi/devin/write/DevinLogger.kt
@@ -1,11 +1,0 @@
-package com.khosravi.devin.write
-
-interface DevinLogger {
-
-    fun doIfEnable(action: (logger: DevinLogger) -> Unit)
-    fun log(message: String)
-    fun log(tag: String, message: String = "")
-    fun logCallerFunc()
-    fun logCallerFunc(tag: String? = null, message: String? = null, enableParentName: Boolean = true)
-}
-
