+++ devin-no-op/src/main/java/com/khosravi/devin/write/DevinTool.kt
@@ -1,6 +1,7 @@
 package com.khosravi.devin.write
 
 import android.content.Context
+import io.nasser.devin.api.DevinLogger
 
 class DevinTool private constructor(
     val logger: DevinLogger?
